move navbar to components
